fix(core): move back to task-orchestrator for detection
